Fixed a bug where adding maze files to the queue that weren't in the same folder as the application wouldn't actually run them. Maze queue files now store relative paths to the added maze files which should make using them unmodified on other computers very easy.
Added a "Cancel" button alongside "Run Trials" and "Test Map." The "Run Trials" button is disabled if there is nothing in the queue. Error message boxes will popup if an enqueued maze was unable to load, but it will continue in the queue.

TODO: use a QActionGroup with mutual exclusion for the 3 buttons.